split lib/orocos/ports.rb into single-class files and refactor the tests
The tests have been rewritten to use Ruby tasks, which make them that
much faster.
