+++ lib/orocos.rb
@@ -48,7 +48,9 @@ require 'orocos/name_service'
 require 'orocos/task_context_base'
 require 'orocos/task_context'
 require 'orocos/ports_base'
-require 'orocos/ports'
+require 'orocos/port'
+require 'orocos/input_port'
+require 'orocos/output_port'
 require 'orocos/operations'
 require 'orocos/process'
 require 'orocos/corba'
